BUG: fix for gcc 341

+++ Common/KWCommon/ProcessXML.cxx
@@ -52,8 +52,8 @@ public:
 
   void CheckSplit(const char* title, const char* file, int force=0)
     {
-    if ( (this->Stream.tellp() - this->CurrentPosition) > this->MaxLen ||
-      force )
+    if ( (static_cast<long>(this->Stream.tellp()) - this->CurrentPosition) > 
+         this->MaxLen || force )
       {
       this->Count ++;
       this->Stream << ";" << endl;
